[compiler] Fix the waiting issue.
In case there was an incomplete type in the function call argument the
compiler was waiting for incorrect ID, so the call instruction was never
analyzed. This was fixed by adding another argument to the function
checking the type completness, set to actual incomplete sub-type in case
one was found. The compiler can later wait for this type ID.